Solvetado bug en validaciones de request que no exitían y separando modelo base de ORM
