+++ app/Models/User.php
@@ -24,6 +24,7 @@ class User extends Model
             'name' => 'required|max:255',
             'email' => 'required|unique:'.$this->table.','.$this->id.'|max:255',
             'password' => 'required|min:6',
+            'newpassword' => 'min:6',
         ];
     }
 }
