+++ config/database.php
@@ -4,6 +4,16 @@ return [
     
     /*
     |--------------------------------------------------------------------------
+    | Clave primaria
+    |--------------------------------------------------------------------------
+    |
+    | Indica el nombre de la clave primaria en tu BBDD
+    | 
+    */
+    'primaryKey' => 'id',
+    
+    /*
+    |--------------------------------------------------------------------------
     | Conexión por defecto
     |--------------------------------------------------------------------------
     |
